[POAE7-2399] [Integration] unit test for inner join support
