+++ cider-velox/test/CMakeLists.txt
@@ -175,9 +175,9 @@ add_executable(CiderOperatorAggOpTest CiderOperatorAggOpTest.cpp)
 target_link_libraries(CiderOperatorAggOpTest ${VELOX_PLUGIN_TEST_LIBS})
 add_test(CiderOperatorAggOpTest CiderOperatorAggOpTest ${TEST_ARGS})
 
-add_executable(CiderHashJoinTest CiderHashJoinTest.cpp)
-target_link_libraries(CiderHashJoinTest ${VELOX_PLUGIN_TEST_LIBS})
-add_test(CiderHashJoinTest CiderHashJoinTest ${TEST_ARGS})
+add_executable(CiderOperatorHashJoinTest CiderOperatorHashJoinTest.cpp)
+target_link_libraries(CiderOperatorHashJoinTest ${VELOX_PLUGIN_TEST_LIBS})
+add_test(CiderOperatorHashJoinTest CiderOperatorHashJoinTest ${TEST_ARGS})
 
 
 add_subdirectory(expression)
